Add filesystem to test the ls and cd commands
A small filesystem tree is added that can be used to test the FTP
server. This contains a directory with two directories and a file, a
directory with a file, a directory without a file, a file with some
data and an empty file. The FTP server shall serve these files and
directories properly.